Added a spotless:check and formatted code accordingly.

+++ src/main/java/com/intel/qat/Native.java
@@ -14,25 +14,22 @@ import java.nio.file.Files;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
 
-/**
- * Class for loading system library - libqat-java.so
- */
+/** Class for loading system library - libqat-java.so */
 class Native {
   private static boolean loaded = false;
   private static String extension = "";
 
   static boolean isLoaded() {
-    if (loaded)
-      return true;
+    if (loaded) return true;
     try {
-      AccessController.doPrivileged(new PrivilegedAction<Void>() {
-        public Void run() {
-          System.loadLibrary(
-              "qat-java"); // Could not load native library from
-                           // "java.library.path", try loading from jar
-          return null;
-        }
-      });
+      AccessController.doPrivileged(
+          new PrivilegedAction<Void>() {
+            public Void run() {
+              System.loadLibrary("qat-java"); // Could not load native library from
+              // "java.library.path", try loading from jar
+              return null;
+            }
+          });
 
       loaded = true;
     } catch (UnsatisfiedLinkError e) {
@@ -42,8 +39,7 @@ class Native {
   }
 
   static String getLibName() {
-    return "/com/intel/qat/" + getOSName() + "/" + getOSArch() + "/"
-        + "libqat-java" + extension;
+    return "/com/intel/qat/" + getOSName() + "/" + getOSArch() + "/" + "libqat-java" + extension;
   }
 
   static String getOSName() {
@@ -52,9 +48,7 @@ class Native {
     if (os.contains("Linux")) {
       ret = "linux";
       extension = ".so";
-    } else
-      throw new UnsupportedOperationException(
-          "Operating System is not supported");
+    } else throw new UnsupportedOperationException("Operating System is not supported");
     return ret;
   }
 
@@ -63,31 +57,26 @@ class Native {
   }
 
   static synchronized void loadLibrary() {
-    if (isLoaded())
-      return;
+    if (isLoaded()) return;
     String libName = getLibName();
     File tempNativeLib = null;
     File tempNativeLibLock = null;
     try (InputStream in = Native.class.getResourceAsStream(libName)) {
       if (in == null) {
         throw new UnsupportedOperationException(
-            "Unsupported OS/arch, cannot find " + libName
-            + ". Please try building from source.");
+            "Unsupported OS/arch, cannot find " + libName + ". Please try building from source.");
       }
       // To avoid race condition with other concurrently running Java processes
       // using qat-java create the .lck file first.
 
-      tempNativeLibLock =
-          File.createTempFile("libqat-java", extension + ".lck");
-      tempNativeLib = new File(
-          tempNativeLibLock.getAbsolutePath().replaceFirst(".lck$", ""));
+      tempNativeLibLock = File.createTempFile("libqat-java", extension + ".lck");
+      tempNativeLib = new File(tempNativeLibLock.getAbsolutePath().replaceFirst(".lck$", ""));
       try (FileOutputStream out = new FileOutputStream(tempNativeLib)) {
         byte[] buf = new byte[4096];
         int bytesRead;
         while (true) {
           bytesRead = in.read(buf);
-          if (bytesRead == -1)
-            break;
+          if (bytesRead == -1) break;
           out.write(buf, 0, bytesRead);
         }
       }
@@ -96,21 +85,19 @@ class Native {
         throw new IOException("Failed to load native qat-java library");
       }
       File finalTempNativeLib = tempNativeLib;
-      AccessController.doPrivileged(new PrivilegedAction<Void>() {
-        public Void run() {
-          System.load(finalTempNativeLib.getAbsolutePath());
-          return null;
-        }
-      });
+      AccessController.doPrivileged(
+          new PrivilegedAction<Void>() {
+            public Void run() {
+              System.load(finalTempNativeLib.getAbsolutePath());
+              return null;
+            }
+          });
       loaded = true;
     } catch (IOException e) {
-      throw new ExceptionInInitializerError(
-          "Failed to load native qat-java library");
+      throw new ExceptionInInitializerError("Failed to load native qat-java library");
     } finally {
-      if (tempNativeLib != null)
-        tempNativeLib.deleteOnExit();
-      if (tempNativeLibLock != null)
-        tempNativeLibLock.deleteOnExit();
+      if (tempNativeLib != null) tempNativeLib.deleteOnExit();
+      if (tempNativeLibLock != null) tempNativeLibLock.deleteOnExit();
     }
   }
 }
